Move hystrimp/example/ to example/
