+++ example/client.go
