+++ example/remote.go
